feat: enhance IntegrityImpactWidget and Dashboard to conditionally render TechnicalDetailsWidget with additional security level props for improved clarity and functionality; update SecurityResourcesWidget tests to include new props and additional test cases
