+++ src/components/widgets/SecurityResourcesWidget.test.tsx
@@ -5,13 +5,27 @@ import { SECURITY_LEVELS } from "../../constants/appConstants";
 
 describe("SecurityResourcesWidget", () => {
   it("renders without crashing", () => {
-    render(<SecurityResourcesWidget securityLevel="None" />);
+    render(
+      <SecurityResourcesWidget
+        securityLevel="None"
+        availabilityLevel="None"
+        integrityLevel="None"
+        confidentialityLevel="None"
+      />
+    );
     expect(screen.getByTestId("widget-security-resources")).toBeInTheDocument();
     expect(screen.getByText("Security Resources")).toBeInTheDocument();
   });
 
   it("displays resources relevant to the security level", () => {
-    render(<SecurityResourcesWidget securityLevel="High" />);
+    render(
+      <SecurityResourcesWidget
+        securityLevel="High"
+        availabilityLevel="High"
+        integrityLevel="High"
+        confidentialityLevel="High"
+      />
+    );
     expect(screen.getByText("Documentation")).toBeInTheDocument();
     expect(
       screen.getByText("Security Implementation Guide")
@@ -23,7 +37,14 @@ describe("SecurityResourcesWidget", () => {
   });
 
   it("displays security training resources", () => {
-    render(<SecurityResourcesWidget securityLevel="High" />);
+    render(
+      <SecurityResourcesWidget
+        securityLevel="High"
+        availabilityLevel="High"
+        integrityLevel="High"
+        confidentialityLevel="High"
+      />
+    );
     expect(screen.getByText("Security Training")).toBeInTheDocument();
     expect(screen.getByText("Security Awareness Training")).toBeInTheDocument();
     expect(
@@ -32,11 +53,42 @@ describe("SecurityResourcesWidget", () => {
   });
 
   it("displays external resources", () => {
-    render(<SecurityResourcesWidget securityLevel="High" />);
+    render(
+      <SecurityResourcesWidget
+        securityLevel="High"
+        availabilityLevel="High"
+        integrityLevel="High"
+        confidentialityLevel="High"
+      />
+    );
     expect(screen.getByText("External Resources")).toBeInTheDocument();
     expect(screen.getByText("Industry Security Standards")).toBeInTheDocument();
     expect(
       screen.getByText("Regulatory Compliance Guides")
     ).toBeInTheDocument();
+  });
+
+  it("renders resource links correctly", () => {
+    render(
+      <SecurityResourcesWidget
+        securityLevel="High"
+        availabilityLevel="High"
+        integrityLevel="High"
+        confidentialityLevel="High"
+      />
+    );
+    // ...existing assertions...
+  });
+
+  it("handles different security levels", () => {
+    render(
+      <SecurityResourcesWidget
+        securityLevel="High"
+        availabilityLevel="Moderate"
+        integrityLevel="High"
+        confidentialityLevel="Very High"
+      />
+    );
+    // ...existing assertions...
   });
 });
